Revert "marked-md components"
This reverts commit cfd0e8370137288216673761dbab3a9fb665bb33.
