--- src/lib/components/flowbite/marked/components/MarkdownBr.svelte
@@ -1,10 +1,0 @@
-<script lang="ts">
-import type { Tokens } from 'marked'
-import type { MarkdownOptions, Renderers } from '../markedConfiguration'
-
-export const token: Tokens.Br = undefined
-export const options: MarkdownOptions = undefined
-export const renderers: Renderers = undefined
-</script>
-
-<br />
