--- src/lib/components/flowbite/marked/components/index.ts
@@ -1,19 +1,0 @@
-export { default as MarkdownHeading } from './MarkdownHeading.svelte';
-export { default as BlockQuote } from './BlockQuote.svelte';
-export { default as MarkdownList } from './list/MarkdownList.svelte';
-export { default as MarkdownListItem } from './list/MarkdownListItem.svelte';
-export { default as MarkdownBr } from './MarkdownBr.svelte';
-export { default as MarkdownCode } from './code/MarkdownCode.svelte';
-export { default as MarkdownCodeSpan } from './code/MarkdownCodeSpan.svelte';
-export { default as MarkdownTable } from './table/MarkdownTable.svelte';
-export { default as MarkdownHtml } from './MarkdownHtml.svelte';
-export { default as MarkdownParagraph } from './MarkdownParagraph.svelte';
-export { default as MarkdownLink } from './MarkdownLink.svelte';
-export { default as MarkdownText } from './MarkdownText.svelte';
-export { default as MarkdownDfn } from './MarkdownDfn.svelte';
-export { default as MarkdownDel } from './MarkdownDel.svelte';
-export { default as MarkdownEm } from './Em.svelte';
-export { default as MarkdownHr } from './MarkdownHr.svelte';
-export { default as MarkdownStrong } from './MarkdownStrong.svelte';
-export { default as MarkdownImage } from './MarkdownImage.svelte';
-export { default as MarkdownSpace } from './MarkdownSpace.svelte';
