--- src/lib/components/flowbite/marked/suppressWarnings.ts
@@ -1,15 +1,0 @@
-import { onMount } from 'svelte'
-
-export function suppressWarnings() {
-  const origWarn = console.warn
-
-  console.warn = (message: string) => {
-    if (message.includes('unknown prop')) return
-    if (message.includes('unexpected slot')) return
-    origWarn(message)
-  }
-
-  onMount(() => {
-    console.warn = origWarn
-  })
-}
