--- src/lib/components/flowbite/marked/typeography/Dfn.svelte
@@ -1,16 +1,0 @@
-<script lang="ts">
-	import type { Tokens } from 'marked';
-	import {
-		MarkdownTokens,
-		type Renderers,
-		type MarkdownOptions
-	} from '@magidoc/plugin-svelte-marked';
-
-	export const token: Tokens.Blockquote = undefined;
-	export const options: MarkdownOptions = undefined;
-	export const renderers: Renderers = undefined;
-</script>
-
-<dfn>
-	<slot />
-</dfn>
