+++ src/lib/components/flowbite/marked/typeography/Link.svelte
@@ -2,14 +2,13 @@
 	import type { Token, Tokens } from 'marked';
 	import {
 		MarkdownTokens,
-		urlUtils,
 		type MarkdownOptions,
 		type Renderers
 	} from '@magidoc/plugin-svelte-marked';
 	import { A, type PweightType } from 'flowbite-svelte';
 	import type { ComponentProps } from 'svelte';
 	import { twMerge } from 'tailwind-merge';
-	const { isRelative, joinUrlPaths } = urlUtils;
+
 	export let token: Tokens.Link &
 		ComponentProps<A> &
 		Tokens.Generic & {
@@ -35,12 +34,3 @@
 <A href={token.href}>
 	<MarkdownTokens tokens={token.tokens} {renderers} {options} />
 </A>
-
-<!-- <a
-  href={isRelative(token.href) 
-    ? joinUrlPaths(options.baseUrl, token.href)
-    : token.href}
-  title={token.title}
->
-  <slot />
-</a> -->
