+++ src/lib/components/flowbite/marked/typeography/List.svelte
@@ -1,20 +1,14 @@
 <script lang="ts">
-	// import MarkdownToken from '$lib/markdown/MarkdownToken.svelte';
+	import type { Token, Tokens } from 'marked';
 	import {
 		MarkdownToken,
+		MarkdownTokens,
 		type MarkdownOptions,
 		type Renderers
 	} from '@magidoc/plugin-svelte-marked';
-	import type { Tokens } from 'marked';
-	import type { ComponentProps } from 'svelte';
 	import { List } from 'flowbite-svelte';
-
-	// export let token: Tokens.List
-	// export let options: MarkdownOptions
-	// export let renderers: Renderers
-
-	// let component: 'ol' | 'ul'
-	// $: component = token.ordered ? 'ol' : 'ul'
+	import type { ComponentProps } from 'svelte';
+	import type { HeadingLevel } from '$components/blog/ui/card';
 
 	type ListType = 'ul' | 'ol' | 'dl';
 	type ListStyle = 'disc' | 'none' | 'decimal';
@@ -27,13 +21,9 @@
 
 	let tag: ListType = token.ordered ? 'ol' : 'ul';
 	let className = token.nested ? 'ps-5 mt-2 space-y-1' : undefined;
-</script>
 
-<!-- <svelte:element this={component} start={token.start || 1}>
-  {#each token.items as item}
-    <MarkdownToken token={{ ...item }} {options} {renderers} />
-  {/each}
-</svelte:element> -->
+	// $: console.log('List', { token });
+</script>
 
 <List bind:tag start={token.start} class={className}>
 	{#each token.items as item}
