+++ src/lib/components/flowbite/marked/typeography/ListItem.svelte
@@ -1,10 +1,12 @@
 <script lang="ts">
-	import type { MarkdownOptions, Renderers } from '../../markedConfiguration';
-
 	import type { Token, Tokens } from 'marked';
+	import {
+		MarkdownTokens,
+		type MarkdownOptions,
+		type Renderers
+	} from '@magidoc/plugin-svelte-marked';
 	import { Li } from 'flowbite-svelte';
 	import type { ComponentProps } from 'svelte';
-	import { MarkdownTokens } from '@magidoc/plugin-svelte-marked';
 
 	type ListType = 'ul' | 'ol' | 'dl';
 	type ListStyle = 'disc' | 'none' | 'decimal';
@@ -20,12 +22,8 @@
 	}) as Tokens.Generic[];
 	let tag: ListType;
 
-	// export const token: Tokens.ListItem = undefined
-	// export const options: MarkdownOptions = undefined
-	// export const renderers: Renderers = undefined
+	$: console.log('ListItem', { isNested, token });
 </script>
-
-<!-- <li><slot /></li> -->
 
 <Li bind:tag>
 	{token.text}
@@ -35,5 +33,4 @@
 	{:else}
 		{token.text}
 	{/if}
-	<!-- <slot /> -->
 </Li>
