--- src/lib/components/flowbite/marked/typeography/index.ts
@@ -1,19 +1,0 @@
-export { default as Heading } from './Heading.svelte';
-export { default as BlockQuote } from './BlockQuote.svelte';
-export { default as MarkdownList } from './list/MarkdownList.svelte';
-export { default as MarkdownListItem } from './list/MarkdownListItem.svelte';
-export { default as MarkdownCode } from './code/MarkdownCode.svelte';
-export { default as MarkdownCodeSpan } from './code/MarkdownCodeSpan.svelte';
-export { default as Br } from './Br.svelte';
-export { default as Table } from './table/Table.svelte';
-export { default as Html } from './Html.svelte';
-export { default as Paragraph } from './Paragraph.svelte';
-export { default as Link } from './Link.svelte';
-export { default as Text } from './Text.svelte';
-export { default as Dfn } from './Dfn.svelte';
-export { default as Del } from './Del.svelte';
-export { default as Em } from './Em.svelte';
-export { default as Hr } from './Hr.svelte';
-export { default as Strong } from './Strong.svelte';
-export { default as Img } from './Img.svelte';
-export { default as Space } from './Space.svelte';
